Place verbose function on server/proxy.
Previously it was global, causing `go test -race` to fail.
