+++ server.go
@@ -60,42 +60,6 @@ var profileStartOnce = sync.Once{}
 
 var log = logrus.StandardLogger()
 
-type EntryLogger func(entry *logrus.Entry, args ...interface{})
-
-// verbose logs to some level on entry, possibly entry.Info.
-var verbose EntryLogger
-
-// makeLevelLogger returns a function that logs level on entry, possible entry.Info.
-func makeLevelLogger(level logrus.Level) EntryLogger {
-	switch level {
-	case logrus.PanicLevel:
-		return func(entry *logrus.Entry, args ...interface{}) { entry.Panic(args) }
-	case logrus.FatalLevel:
-		return func(entry *logrus.Entry, args ...interface{}) { entry.Fatal(args) }
-	case logrus.ErrorLevel:
-		return func(entry *logrus.Entry, args ...interface{}) { entry.Error(args) }
-	case logrus.WarnLevel:
-		return func(entry *logrus.Entry, args ...interface{}) { entry.Warn(args) }
-	case logrus.InfoLevel:
-		return func(entry *logrus.Entry, args ...interface{}) { entry.Info(args) }
-	case logrus.DebugLevel:
-		return func(entry *logrus.Entry, args ...interface{}) { entry.Debug(args) }
-	default:
-		log.Errorf("Unexpected verbose log level %v (using info)", level)
-		return func(entry *logrus.Entry, args ...interface{}) { entry.Info(args) }
-	}
-}
-
-// getVerboseLogLevel parses logLevelStr or Info if empty.
-func getVerboseLogLevel(logLevelStr string) logrus.Level {
-	if logLevelStr == "" {
-		return logrus.InfoLevel
-	}
-	// ParseLevel logs errors.
-	level, _ := logrus.ParseLevel(logLevelStr)
-	return level
-}
-
 var tracer = trace.GlobalTracer
 
 const defaultTCPReadTimeout = 10 * time.Minute
@@ -144,6 +108,7 @@ type Server struct {
 	pluginMtx sync.Mutex
 
 	enableProfiling bool
+	verbose         EntryLogger
 
 	HistogramAggregates samplers.HistogramAggregates
 
@@ -224,7 +189,7 @@ func NewFromConfig(logger *logrus.Logger, conf Config) (*Server, error) {
 		}
 	}
 
-	verbose = makeLevelLogger(getVerboseLogLevel(conf.VerboseLogLevel))
+	ret.verbose = MakeLevelLogger(GetVerboseLogLevel(conf.VerboseLogLevel))
 
 	if conf.Debug {
 		logger.SetLevel(logrus.DebugLevel)
